Bolsa_de_Trabajo_29042025BT/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-filter==25.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
